Two bug fixes to trashing
The first: No trashing if the no. of IDs in ${OIDs} matches the no. of IDs in .$idlist, even if the folder contains more files.
The second: The dash (-) is an accepted character in the ID, but I didn't include it. So, files and folder with a dash or more in their IDs are constantly trashed and got.