allow downloading multiple files of the same name
Google Drive allows people to have many files having the same name in the same folder! To avoid this, now, gdrive-dl appends a plus sign "+" to the name of the second (before the extension), and two plus signs "++" to the third, and so on.

So, if we have four files called "a.pdf" in the same folder, now, gdrive-dl downloads them as "a.pdf", "a+.pdf", "a++.pdf", and "a+++.pdf". You can download this as a test: https://drive.google.com/folderview?id=0BxnPkEc0tlNcM2l0clNSdDM3LUk