Fix 2 checking and updating bugs
If you used any version from 2017.03.16 to 2017.03.18 (inclusive), please get this version and use it with the `-nf` switch (for the first time only) in every folder you used a buggy version with. Then use it as you wish.

Explanation:
Since 2017.03.16, gdrive-dl supports updating modified files. Because Google Drive now makes the modification date of a folder changes if any of its contents changes, we need to check the folder only and compare its modification date. Thus, gdrive-dl doesn't need to enter a non-modified folder. Here is the problem: the IDs file now contains only the contents of the root folder only! What if a file is moved or renamed? How could we know the old name/path?
The solution was to to copy the data of the contents of the non-modified folder from the old IDs file to the new.

The checking bug was a small bug introduced in 2017.03.17 when `HTTP::Tiny` is replaced with Wget. The problem was that it downloads an HTML file from Google Drive (to check its availability) without ignoring it, so you'd find many HTML files in any folder you run gdrive-dl in.